+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # 0.4.0
 - Added support for projectVersion in KraneFile
+- A Krane project can know detect if it is git inintialized (property IsGitInitialized on KraneProject)
+- Added  Invoke-KraneGitAction
+- Added Get-KraneProjectVersion
 
 # 0.3.0